+++ cql3/statements/create_aggregate_statement.hh
@@ -27,7 +27,7 @@ class create_aggregate_statement final : public create_function_statement_base {
     virtual std::unique_ptr<prepared_statement> prepare(data_dictionary::database db, cql_stats& stats) override;
     future<std::pair<::shared_ptr<cql_transport::event::schema_change>, std::vector<mutation>>> prepare_schema_mutations(query_processor& qp, api::timestamp_type) const override;
 
-    virtual shared_ptr<functions::function> create(query_processor& qp, functions::function* old) const override;
+    virtual shared_ptr<db::functions::function> create(query_processor& qp, db::functions::function* old) const override;
 
     sstring _sfunc;
     shared_ptr<cql3_type::raw> _stype;
