+++ cql3/statements/function_statement.hh
@@ -12,11 +12,13 @@
 #include "cql3/functions/function_name.hh"
 #include "cql3/cql3_type.hh"
 
-namespace cql3 {
-
+namespace db {
 namespace functions {
     class function;
-} // namespace functions
+}
+}
+
+namespace cql3 {
 
 namespace statements {
 
@@ -24,23 +26,23 @@ class function_statement : public schema_altering_statement {
 protected:
     virtual future<> check_access(query_processor& qp, const service::client_state& state) const override;
     virtual void prepare_keyspace(const service::client_state& state) override;
-    functions::function_name _name;
+    db::functions::function_name _name;
     std::vector<shared_ptr<cql3_type::raw>> _raw_arg_types;
     mutable std::vector<data_type> _arg_types;
     static shared_ptr<cql_transport::event::schema_change> create_schema_change(
-            const functions::function& func, bool created);
+            const db::functions::function& func, bool created);
     function_statement(functions::function_name name, std::vector<shared_ptr<cql3_type::raw>> raw_arg_types);
     void create_arg_types(query_processor& qp) const;
     data_type prepare_type(query_processor& qp, cql3_type::raw &t) const;
-    virtual shared_ptr<functions::function> validate_while_executing(query_processor&) const = 0;
+    virtual shared_ptr<db::functions::function> validate_while_executing(query_processor&) const = 0;
 };
 
 // common logic for creating UDF and UDA
 class create_function_statement_base : public function_statement {
 protected:
     virtual void validate(query_processor& qp, const service::client_state& state) const override;
-    virtual shared_ptr<functions::function> create(query_processor& qp, functions::function* old) const = 0;
-    virtual shared_ptr<functions::function> validate_while_executing(query_processor&) const override;
+    virtual shared_ptr<db::functions::function> create(query_processor& qp, db::functions::function* old) const = 0;
+    virtual shared_ptr<db::functions::function> validate_while_executing(query_processor&) const override;
 
     bool _or_replace;
     bool _if_not_exists;
@@ -53,7 +55,7 @@ protected:
 class drop_function_statement_base : public function_statement {
 protected:
     virtual void validate(query_processor&, const service::client_state& state) const override;
-    virtual shared_ptr<functions::function> validate_while_executing(query_processor&) const override;
+    virtual shared_ptr<db::functions::function> validate_while_executing(query_processor&) const override;
 
     bool _args_present;
     bool _if_exists;
